IB/srpt: Increase lid and sm_lid to 32 bits
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-862.el7
commit-author Dasaratharaman Chandramouli <dasaratharaman.chandramouli@intel.com>
commit 4c4736905e456819223c6e879ca8fe7cce877939

srpt contains lid and sm_lid fields which are 16 bits in
length, increase them to 32 bits.

(cherry picked from commit 4c4736905e456819223c6e879ca8fe7cce877939)

+++ drivers/infiniband/ulp/srpt/ib_srpt.h
@@ -323,8 +323,8 @@ struct srpt_port {
 	bool			enabled;
 	u8			port_guid[64];
 	u8			port;
-	u16			sm_lid;
-	u16			lid;
+	u32			sm_lid;
+	u32			lid;
 	union ib_gid		gid;
 	spinlock_t		port_acl_lock;
 	struct work_struct	work;
